models/requirements.txt
# Shadow AI Models Requirements

# Core ML/AI
tensorflow>=2.10.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data processing
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Signal processing
neurokit2>=0.2.0
wfdb>=4.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0

# TinyML deployment
tensorflow-lite>=2.10.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
